Rename `useAnimatedButton` to `useScaleAnimation`
